[send_tx_dart] extract `submitAndWatchExtrinsicWithReport`
